pyproject.toml
[project]
name = "ai-agent-learning"
version = "0.1.0"
description = "LLM을 활용한 AI 에이전트 개발 입문 프로젝트"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11.8,<3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"

